feat : 좋아요 필드 관련 수정 #343

+++ src/main/java/umc/th/juinjang/api/note/shared/service/response/SharedNoteGetResponse.java
@@ -30,7 +30,7 @@ public record SharedNoteGetResponse(
 	String price,
 	String monthlyRent,
 	boolean isLiked,
-	int likedCount,
+	Long likedCount,
 	String period,
 	String updatedAt,
 	Long viewCount,
